P6_complete

+++ back/controllers/sauce.js
@@ -6,11 +6,14 @@ const Sauce = require('../models/sauce');
 // Import file system package //
 //----------------------------//
 const fs = require('fs');
+
+const jwt = require("jsonwebtoken");
 //-------------------------------------//
 // Exports the logic of the POST route //
 //-------------------------------------//
 exports.createSauce = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
   const sauce = new Sauce({
     ...sauceObject,
     imageUrl: `${req.protocol}://${req.get('host')}/images/${
@@ -78,19 +81,63 @@ exports.updateSauce = (req, res, next) => {
 //---------------------------------------//
 // Exports the logic of the DELETE route //
 //---------------------------------------//
-exports.deleteSauce = (req, res, next) => {
-  Sauce.findOne({ _id: req.params.id })
+
+exports.getOneSauce = (req, res, next) => {
+  Sauce.findOne({
+    _id: req.params.id,
+  })
     .then((sauce) => {
-      const filename = sauce.imageUrl.split('/images/')[1];
+      res.status(200).json(sauce);
+    })
+    .catch((error) => {
+      res.status(404).json({
+        error: error,
+      });
+    });
+};
+
+exports.modifySauce = (req, res, next) => {
+  const token = req.headers.authorization.split(" ")[1];
+  const decodedToken = jwt.verify(token, `${process.env.RND_TKN}`);
+  const userId = decodedToken.userId;
+  const sauce = req.body.userId;
+  if (userId == sauce) {
+    const sauceObject = req.file
+      ? {
+          ...JSON.parse(req.body.sauce),
+          imageUrl: `${req.protocol}://${req.get("host")}/images/${
+            req.file.filename
+          }`,
+        }
+      : { ...req.body };
+    Sauce.updateOne(
+      { _id: req.params.id },
+      { ...sauceObject, _id: req.params.id }
+    )
+      .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
+      .catch((error) => res.status(403).json({ error }));
+  } else {
+    res.status(403).json({ message: "Erreur" });
+  }
+};
+
+exports.deleteSauce = (req, res, next) => {
+  const token = req.headers.authorization.split(" ")[1];
+  const decodedToken = jwt.verify(token, `${process.env.RND_TKN}`);
+  const userId = decodedToken.userId;
+  Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+    if (sauce.userId == userId) {
+      const filename = sauce.imageUrl.split("/images/")[1];
       fs.unlink(`images/${filename}`, () => {
         Sauce.deleteOne({ _id: req.params.id })
-          .then(() =>
-            res.status(200).json({ message: 'Sauce deleted successfully.' })
-          )
+          .then(() => res.status(200).json({ message: "Sauce supprimé !" }))
           .catch((error) => res.status(400).json({ error }));
       });
-    })
-    .catch((error) => res.status(500).json({ error }));
+    } else {
+      res.status(403).json({ message: "Erreur" });
+      console.log(sauce.userId);
+    }
+  });
 };
 //-----------------------------------------------//
 // Exports the logic of the GET all sauces route //
